+++ etna/models/deadline_ma.py
@@ -92,7 +92,7 @@ class DeadlineMovingAverageModel(
         return self
 
     def _check_not_used_columns(self, ts: TSDataset):
-        columns = set(ts.columns.get_level_values("feature"))
+        columns = set(ts.features)
         columns_not_used = columns.difference({"target"})
         if columns_not_used:
             warnings.warn(